requirement.txt
Django==2.2
djangorestframework==3.11.0
Pillow==8.1.2
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pytz==2021.1
sqlparse==0.4.1
